Refactor to fix companion objects including non-const declarations
