+++ src/main/kotlin/org/openrndr/plugin/intellij/completion/ColorRGBaCompletionContributor.kt
@@ -13,10 +13,10 @@ import org.jetbrains.kotlin.descriptors.ValueDescriptor
 import org.jetbrains.kotlin.idea.core.completion.DeclarationLookupObject
 import org.jetbrains.kotlin.psi.KtCallExpression
 import org.jetbrains.kotlin.psi.KtDeclarationWithInitializer
-import org.openrndr.plugin.intellij.editor.ColorRGBaColorProvider
-import org.openrndr.plugin.intellij.editor.ColorRGBaColorProvider.Companion.isColorModelPackage
-import org.openrndr.plugin.intellij.editor.ColorRGBaColorProvider.Companion.staticColorMap
 import org.openrndr.plugin.intellij.ui.RoundColorIcon
+import org.openrndr.plugin.intellij.utils.ColorUtil
+import org.openrndr.plugin.intellij.utils.ColorUtil.resolveToColor
+import org.openrndr.plugin.intellij.utils.DescriptorUtil.isColorModelPackage
 
 class ColorRGBaCompletionContributor : CompletionContributor() {
     override fun fillCompletionVariants(parameters: CompletionParameters, result: CompletionResultSet) {
@@ -25,34 +25,32 @@ class ColorRGBaCompletionContributor : CompletionContributor() {
             val lookupObject = element.`object` as? DeclarationLookupObject
             val descriptor = (lookupObject?.descriptor as? ValueDescriptor)?.takeIf { it.isColorModelPackage() }
                 ?: return@runRemainingContributors result.passResult(completionResult)
-            val decoratedLookupElement = decorateLookupElement(element, descriptor)
+            val decoratedLookupElement = element.decorateWithIcon(descriptor)
             result.passResult(completionResult.withLookupElement(decoratedLookupElement))
         }
     }
-
-    private companion object {
-        fun decorateLookupElement(element: LookupElement, descriptor: ValueDescriptor) =
-            object : LookupElementDecorator<LookupElement>(element) {
-                override fun renderElement(presentation: LookupElementPresentation) {
-                    super.renderElement(presentation)
-                    val color = staticColorMap[descriptor.name.identifier] ?: return
-                    presentation.icon = JBUIScale.scaleIcon(RoundColorIcon(color, 16, 14))
-                }
-
-                override fun getExpensiveRenderer() = object : LookupElementRenderer<LookupElement>() {
-                    override fun renderElement(element: LookupElement?, presentation: LookupElementPresentation) {
-                        element?.renderElement(presentation)
-                        val property = element?.psiElement as? KtDeclarationWithInitializer ?: return
-                        val callExpression =
-                            PsiTreeUtil.findChildOfType(property.initializer, KtCallExpression::class.java, false)
-                                ?: return
-                        val leaf = PsiTreeUtil.getDeepestFirst(callExpression)
-                        val color = ColorRGBaColorProvider.getColorFrom(leaf) ?: return
-                        presentation.setTypeText(
-                            presentation.typeText, JBUIScale.scaleIcon(RoundColorIcon(color, 16, 14))
-                        )
-                    }
-                }
-            }
-    }
 }
+
+private fun LookupElement.decorateWithIcon(descriptor: ValueDescriptor) =
+    object : LookupElementDecorator<LookupElement>(this) {
+        override fun renderElement(presentation: LookupElementPresentation) {
+            super.renderElement(presentation)
+            val color = ColorUtil.staticColorMap[descriptor.name.identifier] ?: return
+            presentation.icon = JBUIScale.scaleIcon(RoundColorIcon(color, 16, 14))
+        }
+
+        override fun getExpensiveRenderer() = object : LookupElementRenderer<LookupElement>() {
+            override fun renderElement(element: LookupElement?, presentation: LookupElementPresentation) {
+                element?.renderElement(presentation)
+                val property = element?.psiElement as? KtDeclarationWithInitializer ?: return
+                val callExpression =
+                    PsiTreeUtil.findChildOfType(property.initializer, KtCallExpression::class.java, false)
+                        ?: return
+                val leaf = PsiTreeUtil.getDeepestFirst(callExpression)
+                val color = leaf.resolveToColor() ?: return
+                presentation.setTypeText(
+                    presentation.typeText, JBUIScale.scaleIcon(RoundColorIcon(color, 16, 14))
+                )
+            }
+        }
+    }
